Update package name and class function
